requirements.txt
Django>=6.0,<7
djangorestframework>=3.14
djangorestframework-simplejwt>=5.5
django-cors-headers>=4.9
mysqlclient>=2.2
Pillow>=12.0
